+++ src/components/Spinner.tsx
@@ -4,19 +4,23 @@ import '../styles/css/spinner.css'
 
 export interface SpinnerProps {
   /**
-   * Spinner animation speed
+   * The animation (rotation) speed
    */
   duration?: number
   /**
-   * Spinner display mode
+   * The display mode (color theme)
    */
   mode?: TMode
 }
 
 /**
  * UI component to visualize loading content
+ * @author [CasperSocio](https://github.com/CasperSocio)
+ * @version 0.1.0
+ * @since 0.0.1
+ * @global
  */
-export default function Spinner({
+export function Spinner({
   duration = 1,
   mode = 'default',
   ...props