+++ src/index.ts
@@ -1,13 +1,17 @@
-import Alert from './components/Alert'
-import Button from './components/Button'
-import ButtonGroup from './components/ButtonGroup'
-import Spinner from './components/Spinner'
+import { Alert } from './components/Alert'
+import { Button } from './components/Button'
+import { ButtonGroup } from './components/ButtonGroup'
+import { ContactLink } from './components/ContactLink'
+import { Modal } from './components/Modal'
+import { Spinner } from './components/Spinner'
 import { TMode, TShape } from './typings/types'
 
 export {
   Alert,
   Button,
   ButtonGroup,
+  ContactLink,
+  Modal,
   Spinner
 }